test: split McpToolboxClientImplTest into focused test files for errors, headers, and jsonrpc
